+++ src/components/Breadcrumbs/Breadcrumbs.tsx
@@ -72,7 +72,7 @@ export const Breadcrumbs = React.forwardRef(function Breadcrumbs(
         if (props.showRoot) {
             const item = listItems.shift();
             if (item) {
-                rootWidth = item.scrollWidth;
+                rootWidth = item.offsetWidth;
                 calculatedWidth += rootWidth;
             }
             newVisibleItemsCount++;
@@ -118,8 +118,8 @@ export const Breadcrumbs = React.forwardRef(function Breadcrumbs(
     };
 
     const handleResize = React.useCallback(() => {
-        setCalculated(false);
         setVisibleItemsCount(items.length);
+        setCalculated(false);
     }, [items.length]);
     useResizeObserver({
         ref: listRef,
@@ -134,8 +134,8 @@ export const Breadcrumbs = React.forwardRef(function Breadcrumbs(
     React.useLayoutEffect(() => {
         if (calculated && props.children !== lastChildren.current) {
             lastChildren.current = props.children;
-            setCalculated(false);
             setVisibleItemsCount(items.length);
+            setCalculated(false);
         }
     }, [calculated, items.length, props.children]);