+++ src/components/HelpMark/__stories__/HelpMark.stories.tsx
@@ -2,6 +2,7 @@ import type {Meta, StoryObj} from '@storybook/react-webpack5';
 
 import {Showcase} from '../../../demo/Showcase';
 import {ShowcaseItem} from '../../../demo/ShowcaseItem';
+import {Text} from '../../Text';
 import {HelpMark} from '../HelpMark';
 
 const meta: Meta<typeof HelpMark> = {
@@ -50,6 +51,50 @@ export const Size = {
                 </ShowcaseItem>
                 <ShowcaseItem title="Size xl">
                     <HelpMark {...args} iconSize="xl" />
+                </ShowcaseItem>
+            </Showcase>
+        );
+    },
+} satisfies Story;
+
+export const InsideText = {
+    render: (args) => {
+        return (
+            <Showcase>
+                <ShowcaseItem title="Body 1">
+                    <Text variant="body-1">
+                        Lorem ipsum
+                        <br />
+                        dolor <HelpMark {...args} /> sit
+                        <br />
+                        amet
+                    </Text>
+                </ShowcaseItem>
+                <ShowcaseItem title="Body 2">
+                    <Text variant="body-2">
+                        Lorem ipsum
+                        <br />
+                        dolor <HelpMark {...args} /> sit
+                        <br />
+                        amet
+                    </Text>
+                </ShowcaseItem>
+                <ShowcaseItem title="Body 3">
+                    <Text variant="body-3">
+                        Lorem ipsum
+                        <br />
+                        dolor <HelpMark {...args} /> sit
+                        <br />
+                        amet
+                    </Text>
+                </ShowcaseItem>
+                <ShowcaseItem title="Flex Baseline">
+                    <div style={{display: 'flex', alignItems: 'baseline', gap: 4}}>
+                        <div>Property</div>
+                        <div>......</div>
+                        <div>Value</div>
+                        <HelpMark {...args} />
+                    </div>
                 </ShowcaseItem>
             </Showcase>
         );