+++ src/components/Table/hoc/withTableActions/withTableActions.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 
 import {Ellipsis} from '@gravity-ui/icons';
-import _memoize from 'lodash/memoize';
+import memoize from 'lodash/memoize';
 
 import {useUniqId} from '../../../../hooks';
 import {useBoolean} from '../../../../hooks/private';
@@ -76,6 +76,7 @@ export interface WithTableActionsProps<I> {
     getRowActions?: (item: I, index: number) => TableActionConfig<I>[];
     renderRowActions?: (props: RenderRowActionsProps<I>) => React.ReactNode;
     rowActionsSize?: TableRowActionsSize;
+    rowActionsIcon?: React.ReactNode;
 }
 
 interface WithTableActionsState<I> {
@@ -101,7 +102,7 @@ const DEFAULT_PLACEMENT: PopupPlacement = ['bottom-end', 'top-end'];
 
 type DefaultRowActionsProps<I extends TableDataItem> = Pick<
     WithTableActionsProps<I>,
-    'getRowActions' | 'rowActionsSize'
+    'getRowActions' | 'rowActionsSize' | 'rowActionsIcon'
 > &
     Pick<TableProps<I>, 'isRowDisabled' | 'getRowDescriptor'> & {
         item: I;
@@ -115,6 +116,7 @@ const DefaultRowActions = <I extends TableDataItem>({
     getRowActions,
     getRowDescriptor,
     rowActionsSize,
+    rowActionsIcon,
     isRowDisabled,
     tableQa,
 }: DefaultRowActionsProps<I>) => {
@@ -190,7 +192,7 @@ const DefaultRowActions = <I extends TableDataItem>({
                 aria-expanded={isPopupOpen}
                 aria-controls={rowId}
             >
-                <Icon data={Ellipsis} />
+                {rowActionsIcon ?? <Icon data={Ellipsis} />}
             </Button>
         </div>
     );
@@ -236,6 +238,7 @@ export function withTableActions<I extends TableDataItem, E extends {} = {}>(
                 getRowActions,
                 rowActionsSize,
                 renderRowActions,
+                rowActionsIcon,
                 isRowDisabled,
                 getRowDescriptor,
                 qa,
@@ -251,6 +254,7 @@ export function withTableActions<I extends TableDataItem, E extends {} = {}>(
                     item={item}
                     getRowActions={getRowActions}
                     rowActionsSize={rowActionsSize}
+                    rowActionsIcon={rowActionsIcon}
                     getRowDescriptor={getRowDescriptor}
                     isRowDisabled={isRowDisabled}
                     tableQa={qa}
@@ -259,14 +263,14 @@ export function withTableActions<I extends TableDataItem, E extends {} = {}>(
         };
 
         // eslint-disable-next-line @typescript-eslint/member-ordering
-        private enhanceColumns = _memoize((columns: TableColumnConfig<I>[]) =>
+        private enhanceColumns = memoize((columns: TableColumnConfig<I>[]) =>
             enhanceSystemColumn(columns, (systemColumn) => {
                 systemColumn.template = this.renderBodyCell;
             }),
         );
 
         // eslint-disable-next-line @typescript-eslint/member-ordering
-        private enhanceOnRowClick = _memoize(
+        private enhanceOnRowClick = memoize(
             (
                 onRowClick?: (
                     item: I,