+++ src/hooks/private/useFormResetHandler/types.ts
@@ -1,0 +1,1 @@
+export type UseFormResetHandlerParams = HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement;